refactor(webchat): extract shared DeleteButton (2 copies → 1)
PR-1b of admin refactor. Consolidates the near-identical two-step DeleteButton
duplicated in bots/detail and workspaces/detail.

- new components/admin/delete-button.tsx: generic two-step confirm + inline
  error + optional redirect (internal router); call sites pass resourceName,
  buttonLabel, redirectHref, onDelete
- bots/detail & workspaces/detail: drop local copies (~110 lines), adopt
  shared component, clean now-unused useRouter import

Tracks #765.
